requirements.txt
Django==4.1.7
djangorestframework==3.14.0
pytest==7.2.2
